mejora en lógica de gestión de usuarios y estructura de la api
