+++ react-client/src/pages/User.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
 import TableComponent from '../components/Table/TableComponent';
 import SearchBar from '../components/Search/SearchBar';
 import DeactivateModal from '../components/Modal/DeactivateModal';
@@ -6,7 +6,8 @@ import AddUserModal from '../components/Modal/AddUserModal';
 import EditUserModal from '../components/Modal/EditUserModal';
 import AlertComponent from '../components/Alert/AlertComponent';
 import PaginationComponent from '../components/Pagination/PaginationComponent';
-import { fetchUsers, deactivateUser, addUser, updateUser } from '../api/api';
+import { fetchUsers, deactivateUser, addUser, updateUser } from '../interceptors/UsuarioAPIConexion';
+
 
 const UserPage = () => {
     const [data, setData] = useState([]);
@@ -22,41 +23,68 @@ const UserPage = () => {
 
     const itemsPerPage = 3;
 
-    const loadUsers = async () => {
+    // Función para cargar usuarios desde la API
+    const loadUsers = useCallback(async () => {
         try {
             const users = await fetchUsers();
+            console.log('Usuarios cargados desde la API:', users);
             setData(users);
             setFilteredData(users);
             setTotalPages(Math.ceil(users.length / itemsPerPage));
         } catch (err) {
+            console.error('Error al cargar usuarios:', err);
             setAlert({ show: true, message: err.message, type: 'failure' });
         }
-    };
-
+    }, []);
+    
+    // Cargar usuarios al montar el componente
     useEffect(() => {
         loadUsers();
-    }, []);
+    }, [loadUsers]);
 
+    // Filtrar usuarios en función del término de búsqueda
     useEffect(() => {
-        const filtered = data.filter(user =>
-            (user.nombre || '').toLowerCase().includes(searchTerm.toLowerCase())
-        );
+        console.log('Término de búsqueda:', searchTerm);
+        console.log('Datos originales:', data);
+        
+        const filtered = data.filter(user => {
+            if (!user || (typeof user.name !== 'string' && typeof user.apellido !== 'string' && typeof user.document !== 'number')) {
+                console.log('Usuario inválido:', user);
+                return false;
+            }
+            const fullName = `${user.name} ${user.apellido}`.trim().toLowerCase();
+            const document = user.document.toString();
+            const searchTermLower = searchTerm.trim().toLowerCase();
+            
+            const isMatchName = fullName.includes(searchTermLower);
+            const isMatchDocument = document.includes(searchTermLower);
+            
+            console.log(`Usuario: ${fullName}, Documento: ${document}, Coincide: ${isMatchName || isMatchDocument}`);
+            
+            return isMatchName || isMatchDocument;
+        });
+        
+        console.log('Datos filtrados:', filtered);
+        
         setFilteredData(filtered);
         setTotalPages(Math.ceil(filtered.length / itemsPerPage));
         setCurrentPage(1);
     }, [searchTerm, data]);
 
+    // Manejar el cambio de página
     const handlePageChange = (page) => {
         setCurrentPage(page);
     };
 
-    const handleDeactivateClick = (user) => {
+    // Abrir el modal para desactivar un usuario
+    const handleDeactivateClick = useCallback((user) => {
         if (user && user.id) {
             setSelectedUser(user);
             setIsDeactivateModalOpen(true);
         }
-    };
+    }, []);
 
+    // Confirmar y ejecutar la desactivación de un usuario
     const handleDeactivateConfirm = async () => {
         if (!selectedUser?.id) {
             setAlert({ show: true, message: 'No se ha seleccionado ningún usuario para desactivar.', type: 'failure' });
@@ -75,36 +103,41 @@ const UserPage = () => {
         } catch (err) {
             setAlert({ show: true, message: err.message, type: 'failure' });
         } finally {
-            setIsDeactivateModalOpen(false);
-            setSelectedUser(null);
+            handleCloseModal(); // Llamar a handleCloseModal para actualizar los datos
         }
     };
 
-    const handleAddClick = () => {
+    // Abrir el modal para agregar un nuevo usuario
+    const handleAddClick = useCallback(() => {
         setIsAddUserModalOpen(true);
-    };
+    }, []);
 
+    // Guardar un nuevo usuario y recargar la lista
     const handleSaveUser = async (newUser) => {
         try {
             await addUser(newUser);
             setAlert({ show: true, message: '¡Usuario agregado exitosamente!', type: 'success' });
-            await loadUsers();
+            await loadUsers(); // Recargar la lista de usuarios
         } catch (error) {
             setAlert({ show: true, message: error.message, type: 'failure' });
         } finally {
-            setIsAddUserModalOpen(false);
+            handleCloseModal(); // Llamar a handleCloseModal para actualizar los datos
         }
     };
 
-    const handleEditClick = (user) => {
+    // Abrir el modal para editar un usuario
+    const handleEditClick = useCallback((user) => {
         if (user && user.id) {
             setSelectedUser(user);
             setIsEditUserModalOpen(true);
         }
-    };
+    }, []);
 
+    // Actualizar los datos de un usuario
     const handleUpdateUser = async (updatedData) => {
         try {
+            await updateUser(updatedData.id, updatedData); // Llamada a la API para actualizar el usuario
+            // Actualiza el estado con los datos del usuario actualizado
             const updatedUserList = data.map(user =>
                 user.id === updatedData.id ? { ...user, ...updatedData } : user
             );
@@ -115,11 +148,19 @@ const UserPage = () => {
         } catch (error) {
             setAlert({ show: true, message: error.message, type: 'failure' });
         } finally {
-            setIsEditUserModalOpen(false);
+            handleCloseModal(); // Llamar a handleCloseModal para actualizar los datos
         }
     };
-    
 
+    // Función para cerrar modales y recargar usuarios
+    const handleCloseModal = async () => {
+        setIsDeactivateModalOpen(false);
+        setIsAddUserModalOpen(false);
+        setIsEditUserModalOpen(false);
+        await loadUsers(); // Recargar la lista de usuarios después de cerrar un modal
+    };
+
+    // Calcular los datos de la página actual para la paginación
     const startIndex = (currentPage - 1) * itemsPerPage;
     const paginatedData = filteredData.slice(startIndex, startIndex + itemsPerPage);
 
@@ -158,7 +199,7 @@ const UserPage = () => {
             {isDeactivateModalOpen && (
                 <DeactivateModal
                     isOpen={isDeactivateModalOpen}
-                    onClose={() => setIsDeactivateModalOpen(false)}
+                    onClose={handleCloseModal}
                     onDeactivate={handleDeactivateConfirm}
                     userName={selectedUser?.nombre}
                 />
@@ -166,14 +207,14 @@ const UserPage = () => {
             {isAddUserModalOpen && (
                 <AddUserModal
                     isOpen={isAddUserModalOpen}
-                    onClose={() => setIsAddUserModalOpen(false)}
+                    onClose={handleCloseModal}
                     onSave={handleSaveUser}
                 />
             )}
             {isEditUserModalOpen && (
                 <EditUserModal
                     isOpen={isEditUserModalOpen}
-                    onClose={() => setIsEditUserModalOpen(false)}
+                    onClose={handleCloseModal}
                     onSave={handleUpdateUser}
                     userId={selectedUser?.id}
                 />
